refactor(core); httpCompletion
