+++ kotlin/src/main/kotlin/io/logto/sdk/core/http/HttpCompletion.kt
@@ -1,0 +1,9 @@
+package io.logto.sdk.core.http
+
+interface HttpCompletion<T> {
+    fun onComplete(throwable: Throwable?, result: T?)
+}
+
+interface HttpEmptyCompletion {
+    fun onComplete(throwable: Throwable?)
+}
