--- kotlin/src/main/kotlin/io/logto/sdk/core/http/completion/HttpCompletion.kt
@@ -1,5 +1,0 @@
-package io.logto.sdk.core.http.completion
-
-interface HttpCompletion<T> {
-    fun onComplete(throwable: Throwable?, result: T?)
-}
