--- kotlin/src/main/kotlin/io/logto/sdk/core/http/completion/HttpEmptyCompletion.kt
@@ -1,5 +1,0 @@
-package io.logto.sdk.core.http.completion
-
-interface HttpEmptyCompletion {
-    fun onComplete(throwable: Throwable?)
-}
